pyproject.toml
[project]
name = "energy-demand-mlops"
version = "0.1.0"
description = "Production-grade MLOps pipeline for energy demand forecasting"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "lightgbm>=4.0",
    "mlflow>=2.15",
    "evidently>=0.6",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "pydantic>=2.0",
    "psycopg2-binary>=2.9",
    "sqlalchemy>=2.0",
    "prometheus-client>=0.20",
    "prefect>=3.0",
    "dvc>=3.0",
    "streamlit>=1.30",
    "requests>=2.31",
    "holidays>=0.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.2",
    "mypy>=1.8",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
